Rename index to make use of jest moduleDirectories
