+++ tests/__tests__/vue-router.js
@@ -4,7 +4,7 @@ import App from './components/Router/App.vue'
 import Home from './components/Router/Home.vue'
 import About from './components/Router/About.vue'
 
-import { render, fireEvent } from '../../src'
+import { render, fireEvent } from 'vue-testing-library'
 
 const routes = [
   { path: '/', component: Home },
